training/requirements.txt
# langchain core
langchain>=0.3,<0.4

# Google Gemini integration
google-genai
langchain-google-genai

# Vectorstores / embeddings
sentence-transformers
faiss-cpu
pgvector
psycopg2-binary

# ML / utils
torch
transformers
scikit-learn
pandas
numpy
tqdm
python-dotenv
flask

# pdf
pymupdf